requirements.txt
fastapi[standard]>=0.115.0,<0.120.0
uvicorn[standard]>=0.25.0,<0.27.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
aiokafka>=0.12.0,<0.13.0
httpx>=0.24.0,<0.25.0
loguru==0.7.2
python-dotenv==1.0.1
sshtunnel==0.4.0
paramiko<3.4
openai>=1.45.0,<2.0.0
redis>=5.0.0,<6.0.0
PyYAML>=6.0.0,<7.0.0
